requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
coverage==5.3
distlib==0.3.1
Django==3.1.1
django-webpack-loader==0.7.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
importlib-metadata==1.7.0
install==1.3.3
pipenv==2020.8.13
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.31
virtualenv-clone==0.5.4
zipp==3.1.0
